folder.patch
Fix to the google_folder resource:
(1) Folders should be treated as gone if the state is DELETE_REQUESTED.
This is consistent with how projects are treated in the provider.
(2) Set the 'name' field during the TF import. This is required by the
config-connector CLI, specifically to retain the folder ID when
converting the folder's TF state to a KCC resource skeleton.
(3) Replace 'parent' field with 'parent_org_id' and 'parent_folder_id' so
that we can convert these fields to resource references. This is
necessary since KCC can't map a TF field to the appropriate KRM field
when there are multiple possible KRM fields a TF field can be mapped
back to.
(4) On create, check to see if a folder with the given display_name in
the given parent exists before trying to create a new folder. If one
exists, read that one instead. This allows users to acquire folders
by specifying the folder's display_name and parent in KCC.