deletion_protection_default.patch
The provider is introducing breaking changes to database resources to
prevent accidental stateful resource deletion in Terraform. KCC
customers rely on deletion behavior being consistent with its current
behavior, and the accidental stateful resource deletion scenario
isn't as much of an issue in KCC, so we patch this new `true` default
out.