access_context_manager_access_policy_parent.patch
Replace 'parent' with an 'org_id' field that doesn't require an
"organizations/" prefix. This is done to remove the need for a
valueTemplate for the organization-id container annotation which we
needed previously to add the "organizations/" prefix to the
annotation value. This is done so that we can more easily convert
organization-id annotations to external references by just taking the
annotation value and putting it into `external` without having to
prefix it with "organizations/"