secret_manager_secret_version_destroyed.patch
Change TF read to return non-empty resource when state is DESTROYED.
This was the behavior before the upgrade to TF v4.41.0. This is done
since some users might have been depending on KCC to update
status.destroyTime when the resource is DESTROYED.
TODO(b/254533796): Re-evaluate if this patch is worth keeping.