container_nodepool.patch
Fix to the google_container_node_pool resource:
(1)Add support for importing ContainerNodePool from CAI resource name which uses the legacy 'zones/{zone}' instead
of 'locations/{location}'
(2)Do not require a GET to wait for the node pool to be ready during import.
(3)Add a way to disable autoscaling by setting min_node_count and max_node_count to 0
This change will not be upstreamed to TF (they already allow users to
disable autoscaling by unsetting the field) and was implemented because
Kubernetes does not allow objects to be set from non-null to null.