bigtable_gc_policy_read_empty_on_empty_policy.patch
Return an empty instance state on the TF read if the policy is
"<never>" (i.e. an empty policy). This fixes an issue with
BigtableGCPolicy where it could not create an underlying policy at
all even if it was non-existent. Note that this is a partial fix, and
a fuller fix for BigtableGCPolicy should follow (b/173321896).