tests: Fix shutdown ordering (again)
We need to pass the manager context to the mgr.Start method, not just
the mgr.New method.

Also there may have been some variable shadowing going on, and
possibly some unnecessary waiting on a channel.

Hopefully this will now address timeouts - I started hitting it
locally and this cleaned it up immediately, even with
--test.count=10

Also remove the signal handler in the test; it's not standard and it
breaks immediately with --test.count=10