tests: when creating a folder, inject a wait afterwards
When we create a folder, we then try to reacquire it immediately
afterwards using the eventually-consistent search method.  Eventual
consistency is not good enough, and we regularly see flakes here.

This is a workaround to (hopefully) verify the root cause of these
flakes; we can then move to a strongly-consistent approach in a future
change.

Issue #1199