mockgcp: create tool for openapi -> proto
A few APIs don't have public protos, and rather than change how we
implement our mocks, it is possible to create a proto from the
published OpenAPI documents.

This is a "jig" for creating mocks, and certainly is not in the
"serving path", so there are going to be special cases and incomplete
coverage in this tool.