Backporting https://github.com/GoogleCloudPlatform/magic-modules/pull/9463
splits off the default API endpoint half of HandleSDKDefaults func into its own function and calls that new function after TPC universe domain is set. This allows the basepath defaults to use the new domain if set

The API custom/default endpoints are not used between when HandleSDKDefaults was called to after the TPC universe logic, so this reorder should be safe. HandleSDKDefaults is also not called anywhere else outside of the one use.

provider: fixed an issue where universe domains would not overwrite API endpoints

- [X] Run `make ready-pr` to ensure this PR is ready for review.