WIP: feature: implement "watch" for GCP objects
We can't actually watch, but we can instead track dependencies and
poll them.  By doing this we are able to reduce the number of GCP
requests when multiple KCC resources share a GCP resource and we save
a few kube requests in the common (unchanged) case.  We may also be
able to implement a list optimization in future.

This is really a PoC, it only implements a narrow sliver of
functionality.