Invoke the expander when a composition changes to update resources.
When the composition controller reconciles a change in a composition,
the desired outcome is that affected resources be re-processed to
reflect the changes.
Set up a side-channel to the expander and send events on it when
compositions change. The expander will then enqueue the necessary events
for all relevant resources.