Update composition controller RBAC
Tests all manually grant superuser permissions to the composition
controller for simplicity, but that hid the fact that the controller did
not have the right permissions to carry out its tasks.
This change removes the broad permissions from the tests and samples and
updates the controller RBAC to cover more of the requirements (while
still expecting users to grant any additional permissions as necessary).

As a hack to make preview smoother for customers, also globally grant
superuser to the controller for the short term.