fix: Increase wait time before recreating SQLInstances in drift test
For SQLInstance, the wait time of 10s does not appear to be long enough.
When recreating a resource directly after deleting, the create will fail
with "unknown error", unless waiting for a while longer before
attempting to recreate a SQLInstance directly after deleting.

Previously, SQLInstance was unnecessarily skipping the drift detection
test. Most likely, this was because the recreate wait time was not long
enough. There are probably other resources which could perform the test
also, if we increase their wait times before attempting to recreate.