feat: Add predicate to partially enable SQLInstance direct reconciler
If spec.cloneSource is specified, use direct reconciler. Otherwise, use
the terraform-based reconciler.

However, if `KCC_USE_DIRECT_RECONCILERS=SQLInstance` is set, always use
the direct reconciler.

This commit adds a reusable way for other direct reconcilers to
accomplish a similar goal of partially enabling the direct reconciler,
based on the resource to-be-reconciled. The ReconcileGate can be
registered for any direct reconciler, with custom logic based on the
resource.