Remove incorrect assertions for invalid expander tests
The two tests which verify controller behavior for compositions with
invalid expanders have been asserting that the resultant Plan objects
similarly show an error. This check is incorrect because a resource
relying on an invalid Composition should never generate a Plan to begin
with. An existing bug is creating Plans incorrectly and the test was
mistakenly checking on those.

Remove the check of the Plan and only verify the Composition status.