Add support for deleting resources in reverse order by stage.
This solution makes use of the ApplySet annotations for GroupKinds and
Namespaces. When a facade deletion request is received, the ApplySet
data is read from the Plan to know what GKs to find and delete, and
which namespaces to search. The resources are further filtered by the
ApplySet ID they are automatically labeled with.
Stage order information is added to the Plan so the deletion processing
can iterate over them in reverse.