fix: Correctly perform "Export" in dynamic tests
Previously, the dynamic tests would attempt to use the direct controller
"Export" function if there was any direct controller registered for the
Kind. However, in cases where the direct controller is actively under
development, the Terraform-based controller should be used instead. This
fix modifies the resource context to use the same "reconciler type
detection" logic as the rest of the dynamic test, based on the labels
present in the CRD.