chore: Cleanup duplicate label helper functions
This refactor will cause a slight behavior change when exporting
resources into terraform HCL format. Now, the exported HCL will include
the "managed-by-cnrm" label. I think this is ok, as long as the export
is only used for debugging purposes, and customers do not terraform
apply the exported HCL (and therefore potentially cause confusion when
resource labels incorrectly report they are managed by cnrm when, in
fact, they are not managed by cnrm). It could be argued this new
behavior is "more correct", because the exported HCL now exactly matches
what KCC is applying.