mockgcp: update AlloyDB golden HTTP logs
When I had updated the generated AlloyDB protos to include the outbound public IP fields, it also included a new field  on the cluster level. So we need to update various cluster HTTP golden logs. These aren't really related to outbound public IP but are needed for the tests to pass.