fix: Set resource GVKs in IAM integration tests before reconciling
Previously, the testreconciler could not recognize the IAM resources
because they had empty GVKs. The GVKs were erased by previously calling
kubeClient.Create with the typed resources, which uses the typed client
to decode the result back into the object and erases the GVK.

The dynamic controller tests are immune to this issue because they use
unstructured resources instead of typed resources. Therefore, they use
the unstructured kube client, which decodes the result directly back
into the provided unstructured object, and does not erase the GVK.