fix: Update apigee proto definition to fix naming conflict
When trying to generate protos, got the following error
```
--grpc-gateway_out: ProjectsServer.ProvisionOrganizationProject: project
is a protobuf message type. Protobuf message types cannot be used as
path parameters, use a scalar value type (such as string) instead:
```

It seems to be due to a naming conflict around "project" which is
allowed in the JSON API definition, but causes issues when converting.
Probably the easiest fix is to manually make this update.