feat: Enable direct reconciler for SQLInstance
No longer fall back to TF-based reconciler for resources that do not have
spec.cloneSource set. Instead, use direct reconciler for all resources.

This will help with maintenance / debugging issues going forward. Now,
it is simpler to reason about which controller is active, and which code
path is being used.