feat: Refactor Apigee resource identity and reference code
Deduplicate and simplify the identity logic. Use the parent resource
identity in the child resource identity. Create a general pattern for
building new identities from external string format.

When getting the identity of an existing object, use a method defined on
the object's type. This is a bit more ergonomic. In order to simplify
the code, we no longer check individual fields, so error messages only
show if the identity changes overall.

With this change, it's no longer necessary to refer to all parent
resources in the spec of a child resource. It's only necessary to refer
to the direct parent resource.

Also, updated ApigeeEnvgroupAttachment types to refer only to parent
resource, and removed output field of parent resource id.