Non-determinism alert: This seems to provide better results.
Removing the file name seems to help. Not sure why it was added in the first place.
I was seeing some proto blocks in the generate code.
Fixed the branches.yaml to include correct paths. That removed a lot of erroneous generated code.
But there were still some snippets of proto in the generated code.
Modified the prompts and this seems better.