mockgcp: don't ignore unknown fields/enum values by default
We introduced this so that when the API client had newer fields
than we have in our protos, we would not treat that as an error.

However, this also means that we ignore unknown enum values,
which seems much more problematic.

Let's switch the default to strict mode, we should be OK for
most services!