Add unmanaged fields behavior to BigQueryTable direct controller
If the unmanaged annotation is set, the direct controller is used.
Only support unmanaged fields for policy tags and data policy at the
moment.

Within the Update function, the unmanaged fields are removed before
desired vs actual comparision. The, PATCH is used instead of PUT to
avoid overwriting unmanaged fields.