tests: try to avoid port conflicts
We sometimes see a port conflict on the webhook server.

The logic by which the webhook server picks a port is somewhat fragile,
as it opens a random port but does not actually open it until later.

Moreover, the port selection logic is not exported, so we cannot use it
for mockgcp - it has some (fragile) logic to avoid conflicts.

Instead, have mockgcp try to listen on a different host.